# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amazon Listing Auditor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

/**
 * Either asin or url must be provided
 */
export interface FetchListingBody {
  /** Amazon ASIN (e.g. B09G9FPHY6) */
  asin?: string;
  /** Amazon product URL */
  url?: string;
}

export interface FetchedListing {
  productName: string;
  asin: string;
  /** @nullable */
  category?: string | null;
  title: string;
  bulletPoints: string[];
  imageUrls: string[];
  targetKeywords: string[];
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  rating?: string | null;
}

export interface GeneratedContent {
  title: string;
  bulletPoints: string[];
  keywords: string[];
  htmlDescription: string;
}

export interface GeneratedImages {
  main: string[];
  infographic: string[];
  lifestyle: string[];
}

export interface ScoreDetail {
  /** Score from 0 to 100 */
  score: number;
  issues: string[];
  suggestions: string[];
}

export interface AuditResult {
  titleScore: ScoreDetail;
  bulletScore: ScoreDetail;
  imageScore: ScoreDetail;
  keywordScore: ScoreDetail;
  overallScore: number;
  summary: string;
}

export type AuditStatus = (typeof AuditStatus)[keyof typeof AuditStatus];

export const AuditStatus = {
  pending: "pending",
  complete: "complete",
  failed: "failed",
} as const;

export interface Audit {
  id: number;
  productName: string;
  /** @nullable */
  asin?: string | null;
  /** @nullable */
  category?: string | null;
  overallScore: number;
  status: AuditStatus;
  createdAt: string;
  updatedAt: string;
}

export type AuditWithResultsStatus =
  (typeof AuditWithResultsStatus)[keyof typeof AuditWithResultsStatus];

export const AuditWithResultsStatus = {
  pending: "pending",
  complete: "complete",
  failed: "failed",
} as const;

export interface Competitor {
  id: number;
  auditId: number;
  productName: string;
  /** @nullable */
  asin?: string | null;
  title: string;
  bulletPoints: string[];
  imageCount: number;
  targetKeywords: string[];
  overallScore: number;
  strengths: string[];
  weaknesses: string[];
  createdAt: string;
}

export interface AuditWithResults {
  id: number;
  productName: string;
  /** @nullable */
  asin?: string | null;
  /** @nullable */
  category?: string | null;
  title: string;
  bulletPoints: string[];
  imageUrls: string[];
  targetKeywords: string[];
  overallScore: number;
  status: AuditWithResultsStatus;
  result: AuditResult;
  competitors: Competitor[];
  generatedContent?: GeneratedContent;
  generatedImages?: GeneratedImages;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAuditBody {
  productName: string;
  asin?: string;
  category?: string;
  title: string;
  bulletPoints: string[];
  imageUrls: string[];
  targetKeywords: string[];
}

export interface AddCompetitorBody {
  productName: string;
  asin?: string;
  title: string;
  bulletPoints: string[];
  imageCount: number;
  targetKeywords: string[];
}

export interface AuditStats {
  totalAudits: number;
  averageScore: number;
  highScoreCount: number;
  lowScoreCount: number;
  recentAudits: Audit[];
}