# lib/api-zod/src/generated/types/auditWithResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amazon Listing Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditResult } from "./auditResult";
import type { AuditWithResultsStatus } from "./auditWithResultsStatus";
import type { Competitor } from "./competitor";
import type { GeneratedContent } from "./generatedContent";
import type { GeneratedImages } from "./generatedImages";

export interface AuditWithResults {
  id: number;
  productName: string;
  /** @nullable */
  asin?: string | null;
  /** @nullable */
  category?: string | null;
  title: string;
  bulletPoints: string[];
  imageUrls: string[];
  targetKeywords: string[];
  overallScore: number;
  status: AuditWithResultsStatus;
  result: AuditResult;
  competitors: Competitor[];
  generatedContent?: GeneratedContent;
  generatedImages?: GeneratedImages;
  createdAt: Date;
  updatedAt: Date;
}