# lib/api-zod/src/generated/types/fetchListingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amazon Listing Auditor API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Either asin or url must be provided
 */
export interface FetchListingBody {
  /** Amazon ASIN (e.g. B09G9FPHY6) */
  asin?: string;
  /** Amazon product URL */
  url?: string;
}