# lib/api-zod/src/generated/types/fetchedListing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amazon Listing Auditor API
 * OpenAPI spec version: 0.1.0
 */

export interface FetchedListing {
  productName: string;
  asin: string;
  /** @nullable */
  category?: string | null;
  title: string;
  bulletPoints: string[];
  imageUrls: string[];
  targetKeywords: string[];
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  rating?: string | null;
}