# lib/api-zod/src/generated/types/generatedContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amazon Listing Auditor API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedContent {
  title: string;
  bulletPoints: string[];
  keywords: string[];
  htmlDescription: string;
}